reorganizing functions and tests
