+++ .github/workflows/main.yml
@@ -17,7 +17,7 @@ jobs:
         run: pip install pytest
 
       - name: Run Tests
-        run: pytest client_test.py
+        run: pytest tests
         env:
           PREDICTIONGUARD_API_KEY: ${{ secrets.PREDICTIONGUARD_API_KEY }}
           PREDICTIONGUARD_URL: ${{ secrets.PREDICTIONGUARD_URL }}
