--- client_test.py
@@ -1,617 +1,0 @@
-import os
-import re
-import base64
-
-import pytest
-
-from predictionguard import PredictionGuard
-
-
-#----------------------#
-#    Auth/URL Tests    #
-#----------------------#
-
-
-def test_specified_var():
-    test_api_key = os.environ["PREDICTIONGUARD_API_KEY"]
-    test_url = os.environ["PREDICTIONGUARD_URL"]
-
-    test_client = PredictionGuard(api_key=test_api_key, url=test_url)
-
-    assert test_api_key == test_client.api_key
-    assert test_url == test_client.url
-
-
-def test_env_var():
-    test_client = PredictionGuard()
-
-    assert os.environ["PREDICTIONGUARD_API_KEY"] == test_client.api_key
-    assert os.environ["PREDICTIONGUARD_URL"] == test_client.url
-
-
-def test_fail_var():
-    invalid_key_error = """
-Could not connect to Prediction Guard API with the given api_key. 
-Please check your access api_key and try again.
-""".replace(
-        "\n", ""
-    )
-
-    with pytest.raises(ValueError, match=invalid_key_error):
-        PredictionGuard(api_key="i-will-fail-this-test")
-
-
-def test_fail_url():
-    invalid_url_error = """
-Could not connect to Prediction Guard API with given url. 
-Please check url specified, if no url specified, 
-Please contact support.
-""".replace(
-        "\n", ""
-    )
-
-    with pytest.raises(ValueError, match=invalid_url_error):
-        PredictionGuard(
-            api_key=os.environ["PREDICTIONGUARD_API_KEY"],
-            url="https://www.predictionguard.com/",
-        )
-
-
-#-------------------------#
-#    Completions Tests    #
-#-------------------------#
-
-
-def test_completions_create():
-    test_client = PredictionGuard()
-
-    response = test_client.completions.create(
-        model=os.environ["TEST_MODEL_NAME"], prompt="Tell me a joke"
-    )
-
-    assert len(response["choices"][0]["text"]) > 0
-
-
-def test_completions_create_batch():
-    test_client = PredictionGuard()
-
-    response = test_client.completions.create(
-        model=os.environ["TEST_MODEL_NAME"],
-        prompt=["Tell me a joke.", "Tell me a cool fact."],
-    )
-
-    assert len(response["choices"]) > 1
-    assert len(response["choices"][0]["text"]) > 0
-    assert len(response["choices"][1]["text"]) > 0
-
-
-def test_completions_list_models():
-    test_client = PredictionGuard()
-
-    response = test_client.completions.list_models()
-
-    assert len(response) > 0
-
-
-#------------------#
-#    Chat Tests    #
-#------------------#
-
-
-def test_chat_completions_create():
-    test_client = PredictionGuard()
-
-    response = test_client.chat.completions.create(
-        model=os.environ["TEST_MODEL_NAME"],
-        messages=[
-            {"role": "system", "content": "You are a helpful chatbot."},
-            {"role": "user", "content": "Tell me a joke."},
-        ],
-    )
-
-    assert len(response["choices"][0]["message"]["content"]) > 0
-
-
-def test_chat_completions_create_string():
-    test_client = PredictionGuard()
-
-    response = test_client.chat.completions.create(
-        model=os.environ["TEST_MODEL_NAME"],
-        messages="Tell me a joke"
-    )
-
-    assert len(response["choices"][0]["message"]["content"]) > 0
-
-
-def test_chat_completions_create_stream():
-    test_client = PredictionGuard()
-
-    response_list = []
-    for res in test_client.chat.completions.create(
-        model=os.environ["TEST_MODEL_NAME"],
-        messages=[
-            {"role": "system", "content": "You are a helpful chatbot."},
-            {"role": "user", "content": "Tell me a joke."},
-        ],
-        stream=True,
-    ):
-        response_list.append(res)
-
-    assert len(response_list) > 1
-
-
-def test_chat_completions_create_stream_output_fail():
-    test_client = PredictionGuard()
-
-    streaming_error = "Factuality and toxicity checks are not supported when streaming is enabled.".replace(
-        "\n", ""
-    )
-
-    response_list = []
-    with pytest.raises(ValueError, match=streaming_error):
-        for res in test_client.chat.completions.create(
-            model=os.environ["TEST_MODEL_NAME"],
-            messages=[
-                {"role": "system", "content": "You are a helpful chatbot."},
-                {"role": "user", "content": "Tell me a joke."},
-            ],
-            stream=True,
-            output={"toxicity": True},
-        ):
-            response_list.append(res)
-
-
-def test_chat_completions_create_vision_image_file():
-    test_client = PredictionGuard()
-
-    response = test_client.chat.completions.create(
-        model=os.environ["TEST_VISION_MODEL"],
-        messages=[
-            {
-                "role": "user",
-                "content": [
-                    {"type": "text", "text": "What is in this image?"},
-                    {
-                        "type": "image_url",
-                        "image_url": {"url": "fixtures/test_image1.jpeg"},
-                    },
-                ],
-            }
-        ],
-    )
-
-    assert len(response["choices"][0]["message"]["content"]) > 0
-
-
-def test_chat_completions_create_vision_image_url():
-    test_client = PredictionGuard()
-
-    response = test_client.chat.completions.create(
-        model=os.environ["TEST_VISION_MODEL"],
-        messages=[
-            {
-                "role": "user",
-                "content": [
-                    {"type": "text", "text": "What is in this image?"},
-                    {
-                        "type": "image_url",
-                        "image_url": {
-                            "url": "https://farm4.staticflickr.com/3300/3497460990_11dfb95dd1_z.jpg"
-                        },
-                    },
-                ],
-            }
-        ],
-    )
-
-    assert len(response["choices"][0]["message"]["content"]) > 0
-
-
-def test_chat_completions_create_vision_image_b64():
-    test_client = PredictionGuard()
-
-    with open("fixtures/test_image1.jpeg", "rb") as image_file:
-        b64_image = base64.b64encode(image_file.read()).decode("utf-8")
-
-    response = test_client.chat.completions.create(
-        model=os.environ["TEST_VISION_MODEL"],
-        messages=[
-            {
-                "role": "user",
-                "content": [
-                    {"type": "text", "text": "What is in this image?"},
-                    {"type": "image_url", "image_url": {"url": b64_image}},
-                ],
-            }
-        ],
-    )
-
-    assert len(response["choices"][0]["message"]["content"]) > 0
-
-
-def test_chat_completions_create_vision_data_uri():
-    test_client = PredictionGuard()
-
-    with open("fixtures/test_image1.jpeg", "rb") as image_file:
-        b64_image = base64.b64encode(image_file.read()).decode("utf-8")
-
-    data_uri = "data:image/jpeg;base64," + b64_image
-
-    response = test_client.chat.completions.create(
-        model=os.environ["TEST_VISION_MODEL"],
-        messages=[
-            {
-                "role": "user",
-                "content": [
-                    {"type": "text", "text": "What is in this image?"},
-                    {"type": "image_url", "image_url": {"url": data_uri}},
-                ],
-            }
-        ],
-    )
-
-    assert len(response["choices"][0]["message"]["content"]) > 0
-
-
-def test_chat_completions_create_vision_stream_fail():
-    test_client = PredictionGuard()
-
-    streaming_error = "Streaming is not currently supported when using vision."
-
-    response_list = []
-    with pytest.raises(ValueError, match=streaming_error):
-        for res in test_client.chat.completions.create(
-            model=os.environ["TEST_VISION_MODEL"],
-            messages=[
-                {
-                    "role": "user",
-                    "content": [
-                        {"type": "text", "text": "What is in this image?"},
-                        {
-                            "type": "image_url",
-                            "image_url": {"url": "fixtures/test_image1.jpeg"},
-                        },
-                    ],
-                }
-            ],
-            stream=True,
-        ):
-            response_list.append(res)
-
-
-def test_chat_completions_list_models():
-    test_client = PredictionGuard()
-
-    response = test_client.chat.completions.list_models()
-
-    assert len(response) > 0
-
-
-#------------------------#
-#   Embeddings Tests     #
-#------------------------#
-
-
-def test_embeddings_create_text():
-    test_client = PredictionGuard()
-
-    inputs = "Test embeddings"
-
-    response = test_client.embeddings.create(
-        model=os.environ["TEST_TEXT_EMBEDDINGS_MODEL"], input=inputs
-    )
-
-    assert len(response["data"][0]["embedding"]) > 0
-    assert type(response["data"][0]["embedding"][0]) is float
-
-
-def test_embeddings_create_text_batch():
-    test_client = PredictionGuard()
-
-    inputs = ["Test embeddings", "More test embeddings"]
-
-    response = test_client.embeddings.create(
-        model=os.environ["TEST_TEXT_EMBEDDINGS_MODEL"], input=inputs
-    )
-
-    assert len(response["data"]) > 1
-    assert len(response["data"][0]["embedding"]) > 0
-    assert type(response["data"][0]["embedding"][0]) is float
-    assert len(response["data"][1]["embedding"]) > 0
-    assert type(response["data"][1]["embedding"][0]) is float
-
-
-def test_embeddings_create_text_object():
-    test_client = PredictionGuard()
-
-    inputs = [{"text": "How many computers does it take to screw in a lightbulb?"}]
-
-    response = test_client.embeddings.create(
-        model=os.environ["TEST_MULTIMODAL_EMBEDDINGS_MODEL"], input=inputs
-    )
-
-    assert len(response["data"][0]["embedding"]) > 0
-    assert type(response["data"][0]["embedding"][0]) is float
-
-
-def test_embeddings_create_tokens():
-    test_client = PredictionGuard()
-
-    inputs = [0, 8647, 6, 55720, 59725, 7, 5, 2]
-
-    response = test_client.embeddings.create(
-        model=os.environ["TEST_TEXT_EMBEDDINGS_MODEL"], input=inputs
-    )
-
-    assert len(response["data"][0]["embedding"]) > 0
-    assert type(response["data"][0]["embedding"][0]) is float
-
-
-def test_embeddings_create_tokens_batch():
-    test_client = PredictionGuard()
-
-    inputs = [
-        [0, 8647, 6, 55720, 59725, 7, 5, 2],
-        [0, 5455, 3034, 6, 55720, 59725, 7, 5, 2]
-    ]
-
-    response = test_client.embeddings.create(
-        model=os.environ["TEST_TEXT_EMBEDDINGS_MODEL"], input=inputs
-    )
-
-    assert len(response["data"]) > 1
-    assert len(response["data"][0]["embedding"]) > 0
-    assert type(response["data"][0]["embedding"][0]) is float
-    assert len(response["data"][1]["embedding"]) > 0
-    assert type(response["data"][1]["embedding"][0]) is float
-
-
-def test_embeddings_create_image_file():
-    test_client = PredictionGuard()
-
-    inputs = [{"image": "fixtures/test_image1.jpeg"}]
-
-    response = test_client.embeddings.create(
-        model=os.environ["TEST_MULTIMODAL_EMBEDDINGS_MODEL"], input=inputs
-    )
-
-    assert len(response["data"][0]["embedding"]) > 0
-    assert type(response["data"][0]["embedding"][0]) is float
-
-
-def test_embeddings_create_image_url():
-    test_client = PredictionGuard()
-
-    inputs = [
-        {"image": "https://farm4.staticflickr.com/3300/3497460990_11dfb95dd1_z.jpg"}
-    ]
-
-    response = test_client.embeddings.create(
-        model=os.environ["TEST_MULTIMODAL_EMBEDDINGS_MODEL"], input=inputs
-    )
-
-    assert len(response["data"][0]["embedding"]) > 0
-    assert type(response["data"][0]["embedding"][0]) is float
-
-
-def test_embeddings_create_image_b64():
-    test_client = PredictionGuard()
-
-    with open("fixtures/test_image1.jpeg", "rb") as image_file:
-        b64_image = base64.b64encode(image_file.read()).decode("utf-8")
-
-    inputs = [{"image": b64_image}]
-
-    response = test_client.embeddings.create(
-        model=os.environ["TEST_MULTIMODAL_EMBEDDINGS_MODEL"], input=inputs
-    )
-
-    assert len(response["data"][0]["embedding"]) > 0
-    assert type(response["data"][0]["embedding"][0]) is float
-
-
-def test_embeddings_create_data_uri():
-    test_client = PredictionGuard()
-
-    with open("fixtures/test_image1.jpeg", "rb") as image_file:
-        b64_image = base64.b64encode(image_file.read()).decode("utf-8")
-
-    data_uri = "data:image/jpeg;base64," + b64_image
-
-    inputs = [{"image": data_uri}]
-
-    response = test_client.embeddings.create(
-        model=os.environ["TEST_MULTIMODAL_EMBEDDINGS_MODEL"], input=inputs
-    )
-
-    assert len(response["data"][0]["embedding"]) > 0
-    assert type(response["data"][0]["embedding"][0]) is float
-
-
-def test_embeddings_create_both():
-    test_client = PredictionGuard()
-
-    inputs = [{"text": "Tell me a joke.", "image": "fixtures/test_image1.jpeg"}]
-
-    response = test_client.embeddings.create(
-        model=os.environ["TEST_MULTIMODAL_EMBEDDINGS_MODEL"], input=inputs
-    )
-
-    assert len(response["data"])
-
-
-def test_embeddings_create_text_object_batch():
-    test_client = PredictionGuard()
-
-    inputs = [{"text": "Tell me a joke."}, {"text": "Tell me a fact."}]
-
-    response = test_client.embeddings.create(
-        model=os.environ["TEST_MULTIMODAL_EMBEDDINGS_MODEL"], input=inputs
-    )
-
-    assert len(response["data"]) > 1
-    assert len(response["data"][0]["embedding"]) > 0
-    assert type(response["data"][0]["embedding"][0]) is float
-    assert len(response["data"][1]["embedding"]) > 0
-    assert type(response["data"][1]["embedding"][0]) is float
-
-
-def test_embeddings_create_image_batch():
-    test_client = PredictionGuard()
-
-    inputs = [
-        {"image": "fixtures/test_image1.jpeg"},
-        {"image": "fixtures/test_image2.jpeg"},
-    ]
-
-    response = test_client.embeddings.create(
-        model=os.environ["TEST_MULTIMODAL_EMBEDDINGS_MODEL"], input=inputs
-    )
-
-    assert len(response["data"]) > 1
-    assert len(response["data"][0]["embedding"]) > 0
-    assert type(response["data"][0]["embedding"][0]) is float
-    assert len(response["data"][1]["embedding"]) > 0
-    assert type(response["data"][1]["embedding"][0]) is float
-
-
-def test_embeddings_create_both_batch():
-    test_client = PredictionGuard()
-
-    inputs = [
-        {"text": "Tell me a joke.", "image": "fixtures/test_image1.jpeg"},
-        {"text": "Tell me a fun fact.", "image": "fixtures/test_image2.jpeg"},
-    ]
-
-    response = test_client.embeddings.create(
-        model=os.environ["TEST_MULTIMODAL_EMBEDDINGS_MODEL"], input=inputs
-    )
-
-    assert len(response["data"]) > 1
-    assert len(response["data"][0]["embedding"]) > 0
-    assert type(response["data"][0]["embedding"][0]) is float
-    assert len(response["data"][1]["embedding"]) > 0
-    assert type(response["data"][1]["embedding"][0]) is float
-
-
-def test_embeddings_list_models():
-    test_client = PredictionGuard()
-
-    response = test_client.embeddings.list_models()
-
-    assert len(response) > 0
-
-
-#-----------------------#
-#    Translate Tests    #
-#-----------------------#
-
-
-def test_translate_create():
-    test_client = PredictionGuard()
-
-    response = test_client.translate.create(
-        text="The sky is blue", source_lang="eng", target_lang="fra"
-    )
-
-    assert type(response["best_score"]) is float
-    assert len(response["best_translation"])
-
-
-#------------------------#
-#   Factuality Tests     #
-#------------------------#
-
-
-def test_factuality_check():
-    test_client = PredictionGuard()
-
-    response = test_client.factuality.check(
-        reference="The sky is blue", text="The sky is green"
-    )
-
-    assert type(response["checks"][0]["score"]) is float
-
-
-#----------------------#
-#    Toxicity Tests    #
-#----------------------#
-
-
-def test_toxicity_check():
-    test_client = PredictionGuard()
-
-    response = test_client.toxicity.check(text="This is a perfectly fine statement.")
-
-    assert type(response["checks"][0]["score"]) is float
-
-
-#-----------------#
-#    PII Tests    #
-#-----------------#
-
-
-def test_pii_check():
-    test_client = PredictionGuard()
-
-    response = test_client.pii.check(
-        prompt="Hello my name is John Doe. Please repeat that back to me.",
-        replace=True,
-        replace_method="random",
-    )
-
-    assert len(response["checks"][0]["new_prompt"]) > 0
-
-
-#-----------------------#
-#    Injection Tests    #
-#-----------------------#
-
-
-def test_injection_check():
-    test_client = PredictionGuard()
-
-    response = test_client.injection.check(
-        prompt="hi hello", detect=True
-    )
-
-    assert type(response["checks"][0]["probability"]) is float
-
-
-#-----------------------#
-#    Tokenize Tests     #
-#-----------------------#
-
-
-def test_tokenize_create():
-    test_client = PredictionGuard()
-
-    response = test_client.tokenize.create(
-        model=os.environ["TEST_MODEL_NAME"],
-        input="Tokenize this please."
-    )
-
-    assert len(response) > 0
-    assert type(response[0]["id"]) is int
-
-
-#-----------------------#
-#    No API Key Test    #
-#-----------------------#
-
-
-def test_no_key():
-    del os.environ["PREDICTIONGUARD_API_KEY"]
-
-    no_key_error = """
-No api_key provided or in environment. 
-Please provide the api_key as client = PredictionGuard(api_key=<your_api_key>) 
-or as PREDICTIONGUARD_API_KEY in your environment.
-""".replace(
-        "\n", ""
-    )
-
-    no_key_error = re.escape(no_key_error)
-
-    with pytest.raises(ValueError, match=no_key_error):
-        PredictionGuard()
