+++ predictionguard/__init__.py
@@ -1,6 +1,6 @@
 """Create controlled and compliant AI systems with PredictionGuard."""
 
-from .client import PredictionGuard
+from .client import PredictionGuard as PredictionGuard
 from .version import __version__
 
 __version__ = __version__
